Many updates to code with completed 2023 dataset. Adjusted and tweaked so that everything passes standard issue check, with a few statistics that I may want to go back and bolster code for.
CPUE moved to 4e... we might not want it to stay there, but it.s there for now to get it out of the category 1 statistics for interdisc. outcome papers.

The only thing that still needs updates is the world map - rdgal is no longer a package, and need to change map to sf. other than that, problem.

have a few paper Qs that might adjust these data to ask David and Sam, but themes in these 2023_12_4 maps should stay consistent for the final publication